Move detailed mortgage into new page-level component
